+++ OpenRA.Mods.Common/Traits/Render/WithDecoration.cs
@@ -54,7 +54,7 @@ namespace OpenRA.Mods.Common.Traits.Render
 			return wr.Palette(Info.IsPlayerPalette ? Info.Palette + self.Owner.InternalName : Info.Palette);
 		}
 
-		protected override IEnumerable<IRenderable> RenderDecoration(Actor self, WorldRenderer wr, int2 screenPos)
+		protected override IEnumerable<IRenderable> RenderDecoration(Actor self, WorldRenderer wr, float2 screenPos)
 		{
 			if (anim == null)
 				return [];
