+++ OpenRA.Mods.Common/Traits/Render/WithStoresResourcesPipsDecoration.cs
@@ -80,7 +80,7 @@ namespace OpenRA.Mods.Common.Traits.Render
 			return Info.EmptySequence;
 		}
 
-		protected override IEnumerable<IRenderable> RenderDecoration(Actor self, WorldRenderer wr, int2 screenPos)
+		protected override IEnumerable<IRenderable> RenderDecoration(Actor self, WorldRenderer wr, float2 screenPos)
 		{
 			pips.PlayRepeating(Info.EmptySequence);
